requirements.txt
langgraph
sentence-transformers
langchain_google_genai
chromadb
langfuse
crawl4ai
dotenv
json_repair
streamlit
beautifulsoup4
langchain
pydantic
PyYAML


## DOCS: https://docs.crawl4ai.com/api/parameters/